refactor out make dirs, finalise all refactorings
