+++ datashuttle/datashuttle.py
@@ -179,10 +179,14 @@ class DataShuttle:
         utils.log("\nFormatting Names...")
         ds_logger.log_names(["sub_names", "ses_names"], [sub_names, ses_names])
 
-        sub_names = self._format_names(sub_names, "sub")
+        sub_names = formatting.check_and_format_names(
+            self.cfg, sub_names, "sub"
+        )
 
         if ses_names is not None:
-            ses_names = self._format_names(ses_names, "ses")
+            ses_names = formatting.check_and_format_names(
+                self.cfg, ses_names, "ses"
+            )
 
         ds_logger.log_names(
             ["formatted_sub_names", "formatted_ses_names"],
@@ -279,7 +283,8 @@ class DataShuttle:
         if init_log:
             self._start_log("upload_data")
 
-        self._transfer_sub_ses_data(
+        data_transfer.transfer_sub_ses_data(
+            self.cfg,
             "upload",
             sub_names,
             ses_names,
@@ -310,7 +315,8 @@ class DataShuttle:
         if init_log:
             self._start_log("download_data")
 
-        self._transfer_sub_ses_data(
+        data_transfer.transfer_sub_ses_data(
+            self.cfg,
             "download",
             sub_names,
             ses_names,
@@ -772,405 +778,6 @@ class DataShuttle:
     # Make Directory Trees
     # --------------------------------------------------------------------------------------------------------------------
 
-    def _make_directory_trees(
-        self,
-        sub_names: Union[str, list],
-        ses_names: Union[str, list],
-        data_type: str,
-        log: bool = False,
-    ) -> None:
-        """
-        Entry method to make a full directory tree. It will
-        iterate through all passed subjects, then sessions, then
-        subdirs within a data_type directory. This
-        permits flexible creation of directories (e.g.
-        to make subject only, do not pass session name.
-
-        Ensure sub and ses names are already formatted
-        before use in this function (see _start_log())
-
-        Parameters
-        ----------
-
-        sub_names, ses_names, data_type : see make_sub_dir()
-
-        log : whether to log or not. If True, logging must
-            already be initialised.
-        """
-        if data_type not in [[""], ""]:
-            self._check_data_type_is_valid(data_type, error_on_fail=True)
-
-        for sub in sub_names:
-
-            sub_path = self.cfg.make_path(
-                "local",
-                sub,
-            )
-
-            directories.make_dirs(sub_path, log)
-
-            self._make_data_type_directories(data_type, sub_path, "sub")
-
-            for ses in ses_names:
-
-                ses_path = self.cfg.make_path(
-                    "local",
-                    [sub, ses],
-                )
-
-                directories.make_dirs(ses_path, log)
-
-                self._make_data_type_directories(
-                    data_type, ses_path, "ses", log=log
-                )
-
-    def _make_data_type_directories(
-        self,
-        data_type: Union[list, str],
-        sub_or_ses_level_path: Path,
-        level: str,
-        log: bool = False,
-    ) -> None:
-        """
-        Make data_type folder (e.g. behav) at the sub or ses
-        level. Checks directory_class.Directories attributes,
-        whether the data_type is used and at the current level.
-
-        Parameters
-        ----------
-        data_type : data type (e.g. "behav", "all") to use. Use
-            empty string ("") for none.
-
-        sub_or_ses_level_path : Full path to the subject
-            or session directory where the new directory
-            will be written.
-
-        level : The directory level that the
-            directory will be made at, "sub" or "ses"
-
-        log : whether to log on or not (if True, logging must
-            already be initialised).
-        """
-        if data_type in [[""], ""]:
-            return
-
-        data_type_items = self._get_data_type_items(data_type)
-
-        for data_type_key, data_type_dir in data_type_items:  # type: ignore
-
-            if data_type_dir.used and data_type_dir.level == level:
-
-                data_type_path = sub_or_ses_level_path / data_type_dir.name
-
-                directories.make_dirs(data_type_path, log)
-
-                directories.make_datashuttle_metadata_folder(
-                    data_type_path, log
-                )
-
-    # --------------------------------------------------------------------------------------------------------------------
-    # File Transfer
-    # --------------------------------------------------------------------------------------------------------------------
-
-    def check_transfer_sub_ses_input(
-        self, sub_names, ses_names, data_type
-    ) -> Tuple[List[str], List[str], List[str]]:
-
-        if type(sub_names) == str:
-            sub_names = [sub_names]
-
-        if type(ses_names) == str:
-            ses_names = [ses_names]
-
-        if type(data_type) == str:
-            data_type = [data_type]
-
-        if len(sub_names) > 1 and any(
-            [name in ["all", "all_sub"] for name in sub_names]
-        ):
-            utils.log_and_raise_error(
-                "sub_names must only include 'all' or 'all_subs' if these options are used"
-            )  # TODO: if you pass something that doesn't exist, there is no warning
-
-        if len(ses_names) > 1 and any(
-            [name in ["all", "all_ses"] for name in ses_names]
-        ):
-            utils.log_and_raise_error(
-                "ses_names must only include 'all' or 'all_ses' if these options are used"
-            )
-
-        if len(data_type) > 1 and any(
-            [name in ["all", "all_data_type"] for name in data_type]
-        ):
-            utils.log_and_raise_error(
-                "data_type must only include 'all' or 'all_data_type' if these options are used"
-            )
-
-        return sub_names, ses_names, data_type
-
-    def _transfer_sub_ses_data(
-        self,
-        upload_or_download: str,
-        sub_names: Union[str, List[str]],
-        ses_names: Union[str, List[str]],
-        data_type: str,
-        dry_run: bool,
-        log: bool = True,
-    ) -> None:
-        """
-        Iterate through all data type, sub, ses and transfer directory.
-
-        Parameters
-        ----------
-
-        upload_or_download : "upload" or "download"
-
-        sub_names : see make_sub_dir()
-
-        ses_names : see make_sub_dir()
-
-        data_type : e.g. ephys, behav, histology, funcimg, see make_sub_dir()
-
-        dry_run : see upload_project_dir_or_file()
-        """
-        (
-            sub_names_checked,
-            ses_names_checked,
-            data_type_checked,
-        ) = self.check_transfer_sub_ses_input(sub_names, ses_names, data_type)
-
-        local_or_remote = (
-            "local" if upload_or_download == "upload" else "remote"
-        )
-        base_dir = self.cfg.get_base_dir(local_or_remote)
-
-        # Find sub names to transfer
-        if sub_names_checked in [["all"], ["all_sub"]]:
-            processed_sub_names = directories.search_sub_or_ses_level(
-                self.cfg,
-                base_dir,
-                local_or_remote,
-                search_str=f"{self.cfg.sub_prefix}*",
-            )
-            if sub_names_checked == [
-                "all"
-            ]:  # TODO: just made these list sinitially so can stop al lthese stupid checks
-                processed_sub_names += ["all_non_sub"]
-
-        else:
-            processed_sub_names = self._format_names(
-                sub_names_checked, "sub"
-            )  # TODO: will need to ignore keywords # TODO: save somewhere?
-            processed_sub_names = directories.search_for_wildcards(
-                self, base_dir, local_or_remote, processed_sub_names
-            )
-
-        for sub in processed_sub_names:
-
-            if sub == "all_non_sub":  # all_sub is handled above
-                data_transfer.transfer_all_non_sub_ses_data_type(
-                    self.cfg,
-                    upload_or_download,
-                    local_or_remote,
-                    "all_non_sub",
-                    None,
-                    None,
-                    dry_run,
-                    log,
-                )
-                continue
-
-            self._transfer_data_type(
-                upload_or_download,
-                local_or_remote,
-                data_type_checked,
-                sub,
-                dry_run=dry_run,
-                log=log,
-            )
-
-            # Find ses names  to transfer
-            if ses_names_checked in [
-                ["all"],
-                ["all_ses"],
-            ]:  # TODO: save somewhere?
-                processed_ses_names = directories.search_sub_or_ses_level(
-                    self.cfg,
-                    base_dir,
-                    local_or_remote,
-                    sub,
-                    search_str=f"{self.cfg.ses_prefix}*",
-                )
-                if ses_names_checked == [
-                    "all"
-                ]:  # TODO: just made these list sinitially so can stop al lthese stupid checks
-                    processed_ses_names += ["all_non_ses"]
-            else:
-                processed_ses_names = self._format_names(ses_names, "ses")
-                processed_ses_names = directories.search_for_wildcards(
-                    self,
-                    base_dir,
-                    local_or_remote,
-                    processed_ses_names,
-                    sub=sub,
-                )
-
-            for ses in processed_ses_names:
-
-                if ses == "all_non_ses":  # all_ses is handled above
-                    data_transfer.transfer_all_non_sub_ses_data_type(
-                        self.cfg,
-                        upload_or_download,
-                        local_or_remote,
-                        "all_non_ses",
-                        sub,
-                        None,
-                        dry_run,
-                        log,
-                    )  # TODO: fix
-                    continue
-
-                if any(
-                    [
-                        name in ["all_ses_level_non_data_type", "all"]
-                        for name in data_type_checked
-                    ]
-                ):
-
-                    data_transfer.transfer_all_non_sub_ses_data_type(
-                        self.cfg,
-                        upload_or_download,
-                        local_or_remote,
-                        "all_ses_level_non_data_type",
-                        sub,
-                        ses,
-                        dry_run,
-                        log,
-                    )
-
-                self._transfer_data_type(
-                    upload_or_download,
-                    local_or_remote,
-                    data_type_checked,
-                    sub,
-                    ses,
-                    dry_run=dry_run,
-                    log=log,
-                )
-
-    def _transfer_data_type(
-        self,
-        upload_or_download: str,
-        local_or_remote: str,
-        data_type: List[str],
-        sub: str,
-        ses: Optional[str] = None,
-        dry_run: bool = False,
-        log: bool = False,
-    ) -> None:
-        """
-        Transfer the data_type-level folder at the subject
-        or session level. data_type dirs are got either
-        directly from user input or if "all" is passed searched
-        for in the local / remote directory (for
-        upload / download respectively).
-
-        This can handle both cases of subject level dir (e.g. histology)
-        or session level dir (e.g. ephys).
-
-        Note that the use of upload_or_download / local_or_remote
-        is redundant as the value of local_or_remote is set by
-        upload_or_download, but kept for readability.
-
-        Parameters
-        ----------
-
-        upload_or_download : "upload" or "download"
-
-        local_or_remote : "local" or "remote"
-
-        data_type : e.g. "behav", "all"
-
-        sub : subject name
-
-        ses : Optional session name. If False, directory
-            to transfer will be assumed to be at the
-            subject level.
-
-        dry_run : Show data transfer output but do not
-            actually transfer the data.
-
-        log : Whether to log, if True logging must already
-            be initialized
-        """
-        data_type = [
-            dtype
-            for dtype in data_type
-            if dtype != "all_ses_level_non_data_type"
-        ]
-
-        data_type_items = self._items_from_data_type_input(
-            local_or_remote, data_type, sub, ses
-        )
-
-        level = "ses" if ses else "sub"
-
-        for data_type_key, data_type_dir in data_type_items:  # type: ignore
-
-            if data_type_dir.level == level:
-                if ses:
-                    filepath = os.path.join(sub, ses, data_type_dir.name)
-                else:
-                    filepath = os.path.join(sub, data_type_dir.name)
-
-                data_transfer.move_dir_or_file(
-                    filepath,
-                    self.cfg,
-                    upload_or_download,
-                    dry_run=dry_run,
-                    log=log,
-                )
-
-    def _items_from_data_type_input(
-        self,
-        local_or_remote: str,
-        data_type: Union[list, str],
-        sub: str,
-        ses: Optional[str] = None,
-    ) -> Union[ItemsView, zip]:
-        """
-        Get the list of data_types to transfer, either
-        directly from user input, or by searching
-        what is available if "all" is passed.
-
-        Parameters
-        ----------
-
-        see _transfer_data_type() for parameters.
-        """
-        base_dir = self.cfg.get_base_dir(local_or_remote)
-
-        if data_type not in [
-            "all",
-            ["all"],
-            "all_data_type",
-            ["all_data_type"],
-        ]:  # TODO: make this better
-            data_type_items = self._get_data_type_items(
-                data_type,
-            )
-        else:
-            data_type_items = directories.search_data_dirs_sub_or_ses_level(
-                self,
-                self.cfg.data_type_dirs,
-                base_dir,
-                local_or_remote,
-                sub,
-                ses,
-            )
-
-        return data_type_items
-
     # --------------------------------------------------------------------------------------------------------------------
     # SSH
     # --------------------------------------------------------------------------------------------------------------------
@@ -1187,119 +794,6 @@ class DataShuttle:
     # --------------------------------------------------------------------------------------------------------------------
     # Utils
     # --------------------------------------------------------------------------------------------------------------------
-
-    def _format_names(
-        self, names: Union[list, str], sub_or_ses: str
-    ) -> List[str]:
-        """
-        Format a list of subject or session names, e.g.
-        by ensuring all have sub- or ses- prefix, checking
-        for tags, that names do not include spaces and that
-        there are not duplicates.
-
-        Parameters
-        ----------
-
-        names: str or list containing sub or ses names
-                      (e.g. to make dirs)
-
-        sub_or_ses: "sub" or "ses" - this defines the prefix checks.
-        """
-        if sub_or_ses == "sub":
-            non_sub_names = [
-                "all_ses",
-                "all_non_ses",
-                "all_data_type",
-                "all_ses_level_non_data_type",
-            ]  # TODO: put these somewhere! TEST THIS
-            if any(
-                [names in non_sub_names]
-                + [name in non_sub_names for name in names]
-            ):
-                utils.log_and_raise_error(
-                    f"Cannot use a reserved session or data_type keyword "
-                    f"argument (e.g. {non_sub_names}) "
-                    f"as a subject name"
-                )
-
-        else:
-            non_ses_names = [
-                "all_sub",
-                "all_non_sub",
-                "all_data_type",
-                "all_ses_level_non_data_type",
-            ]  # TODO: put these somewhere!
-            if any(
-                [names in non_ses_names]
-                + [name in non_ses_names for name in names]
-            ):
-                utils.log_and_raise_error(
-                    f"Cannot use a reserved session or data_type "
-                    f"keyword argument (e.g. {non_ses_names}) "
-                    f"as a subject name"
-                )
-
-        prefix = self._get_sub_or_ses_prefix(sub_or_ses)
-        formatted_names = formatting.format_names(names, prefix)
-
-        return formatted_names
-
-    def _get_sub_or_ses_prefix(self, sub_or_ses: str) -> str:
-        """
-        Get the sub / ses prefix (default is "sub-" and "ses-") set in cfgs.
-        These should always be "sub-" or "ses-" by SWC-BIDS.
-        """
-        if sub_or_ses == "sub":
-            prefix = self.cfg.sub_prefix
-        elif sub_or_ses == "ses":
-            prefix = self.cfg.ses_prefix
-        return prefix
-
-    def _check_data_type_is_valid(
-        self, data_type: str, error_on_fail: bool
-    ) -> bool:
-        """
-        Check the passed data_type is valid (must
-        be a key on self.ses_dirs e.g. "behav", or "all")
-        """
-        if type(data_type) == list:
-            valid_keys = list(self.cfg.data_type_dirs.keys()) + ["all"]
-            is_valid = all([type in valid_keys for type in data_type])
-        else:
-            is_valid = (
-                data_type in self.cfg.data_type_dirs.keys()
-                or data_type == "all"
-            )
-
-        if error_on_fail and not is_valid:
-            utils.log_and_raise_error(
-                f"data_type: '{data_type}' "
-                f"is not valid. Must be one of"
-                f" {list(self.cfg.data_type_dirs.keys())}. or 'all'"
-                f" No directories were made."
-            )
-
-        return is_valid
-
-    def _get_data_type_items(
-        self, data_type: Union[str, list]
-    ) -> Union[ItemsView, zip]:
-        """
-        Get the .items() structure of the data type, either all of
-        them (stored in self.cfg.data_type_dirs) or as a single item.
-        """
-        if type(data_type) == str:
-            data_type = [data_type]
-
-        if "all" in data_type:
-            items = self.cfg.data_type_dirs.items()
-        else:
-            items = zip(
-                data_type,
-                [self.cfg.data_type_dirs[key] for key in data_type],
-            )
-
-        return items
 
     def _get_rclone_config_name(
         self, connection_method: Optional[str] = None
